Add missing indexes in tuleap_instances table
The columns `ti_name`, `ti_script_path` and `ti_dbprefix` are used to do
some filtering on the table in pretty standard scenarios.

The table is likely to stay small enough to sustain a full scan without
performance issue but setting indexes will make sure we do not encounter
the issue (or trigger warnings in some DBA performance tools).

Change-Id: I90b68183180d05c032ccfa0412e43c0fabbf7208